# frontend/docs-backend/openapi_QUYNH_v8.0.0.yaml
...
openapi: 3.1.0
info:
  title: Quynh Web API
  description: Cổng thông tin điện tử nội bộ Lữ đoàn Thông tin 21. Xem lịch sử thay
    đổi ở openapi.CHANGELOG.md.
  version: 8.0.0
paths:
...
  /command-threads:
    get:
      tags:
      - command-threads
      summary: List Threads
      operationId: list_threads_command_threads_get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandThreadOut'
                title: Response List Threads Command Threads Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - command-threads
      summary: Create Thread
      operationId: create_thread_command_threads_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandThreadCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}:
    get:
      tags:
      - command-threads
      summary: Get Thread
      operationId: get_thread_command_threads__thread_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/messages:
    post:
      tags:
      - command-threads
      summary: Post Message
      operationId: post_message_command_threads__thread_id__messages_post
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_post_message_command_threads__thread_id__messages_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandMessageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/messages/{message_id}/download:
    get:
      tags:
      - command-threads
      summary: Download Message Attachment
      operationId: download_message_attachment_command_threads__thread_id__messages__message_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: message_id
        in: path
        required: true
        schema:
          type: integer
          title: Message Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/close:
    patch:
      tags:
      - command-threads
      summary: Close Thread
      operationId: close_thread_command_threads__thread_id__close_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommandThreadCloseUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/members:
    post:
      tags:
      - command-threads
      summary: Add Members
      operationId: add_members_command_threads__thread_id__members_post
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberAddRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/members/{user_id}:
    delete:
      tags:
      - command-threads
      summary: Remove Member
      operationId: remove_member_command_threads__thread_id__members__user_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/documents:
    get:
      tags:
      - command-threads
      summary: List Documents
      operationId: list_documents_command_threads__thread_id__documents_get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandThreadDocumentOut'
                title: Response List Documents Command Threads  Thread Id  Documents
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - command-threads
      summary: Upload Document
      operationId: upload_document_command_threads__thread_id__documents_post
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_command_threads__thread_id__documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadDocumentOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/documents/{doc_id}/download:
    get:
      tags:
      - command-threads
      summary: Download Document
      operationId: download_document_command_threads__thread_id__documents__doc_id__download_get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: integer
          title: Doc Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/documents/{doc_id}:
    delete:
      tags:
      - command-threads
      summary: Remove Document
      operationId: remove_document_command_threads__thread_id__documents__doc_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      - name: doc_id
        in: path
        required: true
        schema:
          type: integer
          title: Doc Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/minutes/generate:
    post:
      tags:
      - command-threads
      summary: Generate Minutes
      operationId: generate_minutes_command_threads__thread_id__minutes_generate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandThreadMinutesOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /command-threads/{thread_id}/minutes:
    get:
      tags:
      - command-threads
      summary: List Minutes
      operationId: list_minutes_command_threads__thread_id__minutes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: integer
          title: Thread Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommandThreadMinutesOut'
                title: Response List Minutes Command Threads  Thread Id  Minutes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /chats/{id}/read:
    post:
      tags:
      - chats
      summary: Mark Read
      description: Đánh dấu đã đọc toàn bộ tin nhắn trong cuộc hội thoại.
      operationId: mark_read_chats__id__read_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_post_message_command_threads__thread_id__messages_post:
      properties:
        body:
          type: string
          title: Body
          default: ''
        file:
          anyOf:
          - type: string
            contentMediaType: application/octet-stream
          - type: 'null'
          title: File
      type: object
      title: Body_post_message_command_threads__thread_id__messages_post
...
    Body_upload_document_command_threads__thread_id__documents_post:
      properties:
        title:
          type: string
          title: Title
          default: ''
        visibility:
          type: string
          enum:
          - chung
          - rieng
          title: Visibility
          default: chung
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_document_command_threads__thread_id__documents_post
...
    CommandMessageOut:
      properties:
        id:
          type: integer
          title: Id
        thread_id:
          type: integer
          title: Thread Id
        sender_id:
          type: integer
          title: Sender Id
        sender_full_name:
          type: string
          title: Sender Full Name
        body:
          type: string
          title: Body
        attachment_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Attachment Url
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - thread_id
      - sender_id
      - sender_full_name
      - body
      - attachment_url
      - created_at
      title: CommandMessageOut
    CommandThreadCloseUpdate:
      properties:
        is_closed:
          type: boolean
          title: Is Closed
      type: object
      required:
      - is_closed
      title: CommandThreadCloseUpdate
    CommandThreadCreate:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        member_user_ids:
          items:
            type: integer
          type: array
          title: Member User Ids
      type: object
      required:
      - title
      title: CommandThreadCreate
    CommandThreadDetailOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        classification:
          type: string
          title: Classification
        created_by_id:
          type: integer
          title: Created By Id
        created_by_full_name:
          type: string
          title: Created By Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
        last_message_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Message At
        is_closed:
          type: boolean
          title: Is Closed
        message_count:
          type: integer
          title: Message Count
        unread_count:
          type: integer
          title: Unread Count
        member_count:
          type: integer
          title: Member Count
        messages:
          items:
            $ref: '#/components/schemas/CommandMessageOut'
          type: array
          title: Messages
        members:
          items:
            $ref: '#/components/schemas/MemberOut'
          type: array
          title: Members
      type: object
      required:
      - id
      - title
      - classification
      - created_by_id
      - created_by_full_name
      - created_at
      - last_message_at
      - is_closed
      - message_count
      - unread_count
      - member_count
      - messages
      - members
      title: CommandThreadDetailOut
    CommandThreadDocumentOut:
      properties:
        id:
          type: integer
          title: Id
        thread_id:
          type: integer
          title: Thread Id
        title:
          type: string
          title: Title
        visibility:
          type: string
          enum:
          - chung
          - rieng
          title: Visibility
        file_url:
          type: string
          title: File Url
        file_name:
          type: string
          title: File Name
        file_size:
          type: integer
          title: File Size
        content_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Type
        uploaded_by_id:
          type: integer
          title: Uploaded By Id
        uploaded_by_full_name:
          type: string
          title: Uploaded By Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - thread_id
      - title
      - visibility
      - file_url
      - file_name
      - file_size
      - content_type
      - uploaded_by_id
      - uploaded_by_full_name
      - created_at
      title: CommandThreadDocumentOut
    CommandThreadMinutesOut:
      properties:
        id:
          type: integer
          title: Id
        thread_id:
          type: integer
          title: Thread Id
        content:
          type: string
          title: Content
        message_count:
          type: integer
          title: Message Count
        generated_by_id:
          type: integer
          title: Generated By Id
        generated_by_full_name:
          type: string
          title: Generated By Full Name
        generated_at:
          type: string
          format: date-time
          title: Generated At
      type: object
      required:
      - id
      - thread_id
      - content
      - message_count
      - generated_by_id
      - generated_by_full_name
      - generated_at
      title: CommandThreadMinutesOut
    CommandThreadOut:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        classification:
          type: string
          title: Classification
        created_by_id:
          type: integer
          title: Created By Id
        created_by_full_name:
          type: string
          title: Created By Full Name
        created_at:
          type: string
          format: date-time
          title: Created At
        last_message_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Message At
        is_closed:
          type: boolean
          title: Is Closed
        message_count:
          type: integer
          title: Message Count
        unread_count:
          type: integer
          title: Unread Count
        member_count:
          type: integer
          title: Member Count
      type: object
      required:
      - id
      - title
      - classification
      - created_by_id
      - created_by_full_name
      - created_at
      - last_message_at
      - is_closed
      - message_count
      - unread_count
      - member_count
      title: CommandThreadOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MemberAddRequest:
      properties:
        user_ids:
          items:
            type: integer
          type: array
          minItems: 1
          title: User Ids
      type: object
      required:
      - user_ids
      title: MemberAddRequest
    MemberOut:
      properties:
        user_id:
          type: integer
          title: User Id
        full_name:
          type: string
          title: Full Name
        unit_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Unit Name
        added_at:
          type: string
          format: date-time
          title: Added At
      type: object
      required:
      - user_id
      - full_name
      - added_at
      title: MemberOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer